+++ rdbms/pom.xml
@@ -104,7 +104,7 @@
                 <artifactId>liquibase-maven-plugin</artifactId>
                 <version>4.5.0</version>
                 <configuration>
-                    <changeLogFile>C:\Users\astaf\IdeaProjects\example\Sprint-7\rdbms\src\main\resources\db.changelog-master.yaml</changeLogFile>
+                    <changeLogFile>/src/main/resources/changelog-master.yaml</changeLogFile>
                     <url>jdbc:postgresql://localhost:5432/db</url>
                     <username>postgres</username>
                     <password>postgres</password>
