+++ rdbms/src/main/resources/changelog-master.yaml
@@ -1,0 +1,5 @@
+databaseChangeLog:
+  - include:
+      file: /src/main/resources/init.sql
+
+
