+++ rdbms/src/main/resources/init.sql
@@ -1,15 +1,15 @@
---liquibase formatted sql
-
---changeset rrmasgutov:init
-CREATE TABLE account1
+--changeSet astafex:init
+CREATE TABLE account
 (
     id      BIGSERIAL
         CONSTRAINT account_pk PRIMARY KEY,
-    amount  BIGINT,
+    amount  BIGINT CHECK ( amount > 0 ),
     version BIGINT
-)
-;
+);
 
---changeset astafex:add_index
-CREATE INDEX account1_id_index ON account1 (id)
+--changeSet astafex:add_index
+CREATE INDEX account1_id_index ON account (id);
 
+--changeSet astafex:add_accounts_data
+INSERT INTO account VALUES (1,10000,0);
+INSERT INTO account VALUES (2,20000,0);
